Larger icon for the edit mode
I know this change does not make sense, but I checked it a few weeks ago and forgot to prepare Pull Request.

How to check?
1. Go to edit mode (pencil icon)
2. Set the maximum icons in Preferences> Appearance> Main UI
3. Note that this one and only icon has not grown.

That is why in this unique place we have to load the "big" pencil right away.

It is possible that there is another solution, but I admit that I did not find it. It just works.